Refactored utils, removed lodash.assign
