+++ package.json
@@ -15,6 +15,7 @@
   "scripts": {
     "link-mode": "cross-env NODE_ENV=link webpack --config ./config/webpack.config.js",
     "lint": "eslint src",
+    "lint-fix": "eslint src --fix",
     "start": "cross-env NODE_ENV=development && showroom-scan src && babel-node www/index",
     "test": "mocha src/**/*.spec.js --compilers js:babel-register --require config/test/mocha-setup.js --require ignore-styles",
     "npm-build": "rimraf lib && cross-env NODE_ENV=production WEBPACK_BUNDLE_ANALYZE=true webpack --config ./config/webpack.config.js",
